feat(gallery): v2 with detail view, icons, and 19 chart variants
Gallery now features:
- Sun/moon icon toggle via iced_lucide (replacing text toggler)
- Detail screen with chart + syntax-highlighted source code
- 10 per-type categories (Line, Bar, Pie, Scatter, Area, etc.)
- 19 chart variants including stacked, multi-series, labels, donut
- Action pattern for screen routing (gallery ↔ detail)

Also adds font_variations field for updated iced, cosmic-text patches,
and fixes btc_stream f32→f64 label format signature.

+++ examples/btc_stream/src/main.rs
@@ -499,7 +499,7 @@ async fn fetch_historical_prices(pair: TradingPair) -> Result<Vec<PricePoint>, S
 }
 
 // Helper function to format numbers with thousands separator
-fn currency(value: f32) -> String {
+fn currency(value: f64) -> String {
     let whole = value as i32;
     let s = whole.to_string();
     let mut result = String::new();
